requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-admin-thumbnails==0.2.7
django-ckeditor==6.5.1
django-cleanup==7.0.0
django-cors-headers==4.0.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-mptt-admin==2.5.0
django-rest-swagger==2.2.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pyaml==23.5.9
PyJWT==2.7.0
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==2.0.3
